[artifacts] Enable daily builds
This enables daily artifact builds and slack notifications for the
corresponding pipeline.  Frequency will eventually be changed to on
merge, but I'd like to start with a slower frequency to get a triage
process in place and for a stabilization period.

This also removes the schedule for the Docker context and package tesing
pipelines as these were merged into the artifacts pipeline.  I'll follow
up in a few months to drop the definitions altogether when we're okay
with expiring the logs.